Team04_TermProject/src/mosaicv1/requirements.txt
#torch
#torchvision
#torchaudio
opencv-python-headless
# dlib
dill
face-recognition
ultralytics
matplotlib
facenet-pytorch